requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.12.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cssselect==1.2.0
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
executing==1.2.0
fake-useragent==1.1.1
fastapi==0.89.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gspread==5.7.2
gspread-dataframe==3.3.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.1
jupyter_core==5.1.5
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.26.4
orjson==3.8.5
packaging==23.0
pandas==1.5.3
parse==1.19.0
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
PyDrive==1.3.1
pyee==8.2.2
Pygments==2.14.0
pygsheets==2.0.6
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-html==0.10.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.22.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
w3lib==2.1.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
yarl==1.8.2
zipp==3.11.0
